% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenColors.R
\name{PhenColors}
\alias{PhenColors}
\title{PhenColors}
\usage{
PhenColors(p)
}
\arguments{
\item{p}{Nombre de la Paleta}
}
\description{
Paleta de color Figuras fenologia LabGRS
}
\seealso{
\code{\link{https://drive.google.com/file/d/1STe-xykrX-JQaT8u6gsSpAojIlQI-mxB/view?usp=sharing}}
}
\author{
Matias Olea
}
