% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenColors.R
\name{PhenColors}
\alias{PhenColors}
\title{PhenColors}
\usage{
PhenColors(p)
}
\arguments{
\item{p}{Vector de texto con Nombre de la Paleta p.e. c("Phen3","Phen5","VIloss4","VIloss6,"
VIloss8","VIgain4","VIgain6,"VIgain8","Extreme5","Extreme7","Extreme9","RFD","ColorBlind5","ColorBlind7,"ColorBlind9",
"NaQa","NaNX")}
}
\description{
Paleta de color Figuras fenologia LabGRS
}
\seealso{

}
\author{
Matias Olea
}
