% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenColors.R
\name{PhenColors}
\alias{PhenColors}
\title{PhenColors}
\usage{
PhenColors(p)
}
\arguments{
\item{p}{Vector de texto con nombre de la Paleta p.e. ("Phen3", "Phen5", "VIloss4", "VIloss6,
"VIloss8", "VIgain4", "VIgain6, "VIgain8", "Extreme5", "Extreme7", "Extreme9", "RFD", "ColorBlind5",
"ColorBlind7", "ColorBlind9", "NaQa", "NaNX")}
}
\description{
Paleta de color Figuras fenologia LabGRS
}
\examples{
# Probar la rampa
\donttest{
library(PhenColors)
color <- PhenColors("Extreme5")
barplot(rep(1, length(color)), names.arg = color, col = color)
}
}
\seealso{
\url{https://drive.google.com/file/d/1STe-xykrX-JQaT8u6gsSpAojIlQI-mxB/view}.
}
\author{
Matias Olea
}
